bugfixes, cleanup and refactor utils into modules
