+++ run_deterministic_tests.py
@@ -2,7 +2,7 @@
 This script is used to run deterministic tests on the Fastlane Bot.
 
 The script is run from the command line with the following command:
-`python run_deterministic_tests.py --task <task> --rpc_url <rpc_url> --network <network> --arb_mode <arb_mode>`
+`python run_deterministic_tests.py --task <task> --rpc_url <rpc_url> --network <network> --arb_mode <arb_mode>` --timeout_minutes <timeout_minutes>
 
 The `--task` argument specifies the task to run. The options are:
 - `set_test_state`: Set the test state based on the static_pool_data_testing.csv file.
@@ -21,6 +21,8 @@ The `--arb_mode` argument specifies the arbitrage mode to test. The options are:
 - `multi`: Multi arbitrage mode.
 - `triangle`: Triangle arbitrage mode.
 - `multi_triangle`: Multi triangle arbitrage mode.
+
+The `--timeout_minutes` argument specifies the timeout for the tests (in minutes).
 
 The script uses the `fastlane_bot/tests/deterministic/constants.py` file to get the constants used in the tests.
 
@@ -46,21 +48,16 @@ from web3 import Web3
 from web3.contract import Contract
 
 from fastlane_bot.tests.deterministic.constants import (
-    KNOWN_UNABLE_TO_DELETE,
-    TENDERLY_RPC_KEY,
+    DEFAULT_FROM_BLOCK, KNOWN_UNABLE_TO_DELETE,
+    TENDERLY_RPC_KEY, TEST_FILE_DATA_DIR,
 )
-from fastlane_bot.tests.deterministic.utils import (
-    PoolParamsBuilder,
-    Strategy,
-    StrategyManager,
-    TestPool,
-    Web3Manager,
-    clean_tx_data,
-    get_default_main_args,
-    get_test_strategies,
-    get_tx_data,
-    scan_logs_for_success,
-)
+from fastlane_bot.tests.deterministic.main_args_mock import ArgumentParserMock
+from fastlane_bot.tests.deterministic.test_tx_helper import TestTxHelper
+from fastlane_bot.tests.deterministic.mgr_strategy import StrategyManager
+from fastlane_bot.tests.deterministic.test_pool_params_builder import TestPoolParamsBuilder
+from fastlane_bot.tests.deterministic.test_pool import TestPool
+from fastlane_bot.tests.deterministic.test_strategy import TestStrategy
+from fastlane_bot.tests.deterministic.mgr_web3 import Web3Manager
 
 
 def set_test_state_task(w3: Web3):
@@ -75,7 +72,7 @@ def set_test_state_task(w3: Web3):
     # Import pool data
     # Set the default paths
     static_pool_data_testing_path = os.path.normpath(
-        "fastlane_bot/tests/deterministic/_data/static_pool_data_testing.csv"
+        f"{TEST_FILE_DATA_DIR}/static_pool_data_testing.csv"
     )
 
     test_pools = pd.read_csv(static_pool_data_testing_path, dtype=str)
@@ -84,7 +81,7 @@ def set_test_state_task(w3: Web3):
         for index, test_pools_row in test_pools.iterrows()
     ]
     pools = [pool for pool in pools if pool.is_supported]
-    builder = PoolParamsBuilder(w3)
+    builder = TestPoolParamsBuilder(w3)
 
     # Handle each exchange_type differently for the required updates
     for pool in pools:
@@ -102,7 +99,7 @@ def set_test_state_task(w3: Web3):
 
 
 def get_carbon_strategies_and_delete_task(
-    w3: Web3, carbon_controller: Contract, from_block: int = 1000000
+        w3: Web3, carbon_controller: Contract, from_block: int
 ):
     """
     Get the carbon strategies and delete them.
@@ -136,10 +133,10 @@ def get_carbon_strategies_and_delete_task(
 
 
 def run_tests_on_mode_task(
-    args: argparse.Namespace,
-    w3: Web3,
-    carbon_controller: Contract,
-    test_strategies: Dict,
+        args: argparse.Namespace,
+        w3: Web3,
+        carbon_controller: Contract,
+        test_strategies: Dict,
 ):
     """
     Run tests on the specified arbitrage mode.
@@ -153,17 +150,14 @@ def run_tests_on_mode_task(
     print("\nRunning run_tests_on_mode_task...")
 
     # Initialize the Strategy Manager
-    try:
-        strategy_mgr = StrategyManager(w3, carbon_controller)
-    except Exception as e:
-        print(f"Failed to initialize StrategyManager: {e}")
-        return
+    strategy_mgr = StrategyManager(w3, carbon_controller)
 
     # Get the default main args
-    default_main_args = get_default_main_args()
+    default_main_args = ArgumentParserMock()
     default_main_args.blockchain = args.network
     default_main_args.arb_mode = args.arb_mode
-    default_main_args.timeout = 60 * 4  # 4 minutes
+    default_main_args.timeout = args.timeout
+    default_main_args.rpc_url = args.rpc_url
 
     # Print the default main args
     print(f"\n\n ********\n"
@@ -181,10 +175,10 @@ def run_tests_on_mode_task(
     print(f"strats_created_from_block: {strats_created_from_block}")
 
     # populate a dictionary with all the relevant test strategies
-    test_strategy_txhashs: Dict[Strategy] or Dict = {}
-    for key, args in test_strategies.items():
+    test_strategy_txhashs: Dict[TestStrategy] or Dict = {}
+    for i, (key, args) in enumerate(test_strategies.items()):
         args["w3"] = w3
-        test_strategy = Strategy(**args)
+        test_strategy = TestStrategy(**args)
         test_strategy.get_token_approval(
             token_id=0, approval_address=carbon_controller.address
         )
@@ -192,12 +186,37 @@ def run_tests_on_mode_task(
             token_id=1, approval_address=carbon_controller.address
         )
         tx_hash = strategy_mgr.create_strategy(test_strategy)
-        test_strategy_txhashs[key] = {"txhash": tx_hash} if tx_hash else {}
+        test_strategy_txhashs[str(i + 1)] = {"txhash": tx_hash} if tx_hash else {}
 
     # Write the test strategy txhashs to a file
     test_strategy_txhashs_path = os.path.normpath(
-        "fastlane_bot/tests/deterministic/test_strategy_txhashs.json"
+        f"{TEST_FILE_DATA_DIR}/test_strategy_txhashs.json"
     )
+
+    # Initialize the Strategy Manager
+    strategy_mgr = StrategyManager(w3, carbon_controller)
+
+    # Get the new state of the carbon strategies
+    (
+        strategy_created_df,
+        strategy_deleted_df,
+        remaining_carbon_strategies,
+    ) = strategy_mgr.get_state_of_carbon_strategies(strats_created_from_block)
+
+    new_strats_created = strategy_created_df["id"].to_list()
+    print(f"There have been {len(new_strats_created)} new strategies created")
+
+    print(f"strategy_created_df.columns: {list(strategy_created_df.columns)}")
+    strategy_created_df.to_csv("strategy_created_df.csv")
+
+    print("\nAdd the strategy ids...")
+    for i in test_strategy_txhashs.keys():
+        try:
+            test_strategy_txhashs[i]['strategyid'] = strategy_created_df[
+                strategy_created_df['transaction_hash'] == test_strategy_txhashs[i]['txhash']].id.values[0]
+        except Exception as e:
+            print(f"Add the strategy ids Error: {i}, {e}")
+
     with open(test_strategy_txhashs_path, "w") as f:
         json.dump(test_strategy_txhashs, f)
         f.close()
@@ -209,11 +228,21 @@ def run_results_crosscheck_task():
     """
     print("\nRunning run_results_crosscheck_task...")
 
+    # Initialize the tx helper
+    tx_helper = TestTxHelper()
+
     # Successful transactions on Tenderly are marked by status=1
-    actually_txt_all_successful_txs = scan_logs_for_success()
+    actually_txt_all_successful_txs = tx_helper.tx_scanner()
+    print(f"len(actually_txt_all_successful_txs): {len(actually_txt_all_successful_txs)}")
+
+    # print first 3 examples of actually_txt_all_successful_txs
+    for i, tx in enumerate(actually_txt_all_successful_txs):
+        print(f"\nactually_txt_all_successful_txs[{i}]: {tx}")
+        if i == 2:
+            break
 
     test_results_path = os.path.normpath(
-        "fastlane_bot/data/blockchain_data/ethereum/test_results.json"
+        f"{TEST_FILE_DATA_DIR}/test_results.json"
     )
     with open(test_results_path) as f:
         test_datas = json.load(f)["test_data"]
@@ -221,7 +250,7 @@ def run_results_crosscheck_task():
         print(f"{len(test_datas.keys())} test results imported")
 
     test_strategy_txhashs_path = os.path.normpath(
-        "fastlane_bot/tests/deterministic/test_strategy_txhashs.json"
+        f"{TEST_FILE_DATA_DIR}/test_strategy_txhashs.json"
     )
     with open(test_strategy_txhashs_path) as f:
         test_strategy_txhashs = json.load(f)
@@ -231,14 +260,16 @@ def run_results_crosscheck_task():
     # Loop over the created test strategies and verify test data
     for i in test_strategy_txhashs.keys():
         search_id = test_strategy_txhashs[i]["strategyid"]
-        print(f"Evaluating test {i}, {search_id}")
-        tx_data = get_tx_data(search_id, actually_txt_all_successful_txs)
-        clean_tx_data(tx_data)
+        print(f"Evaluating test {search_id}")
+        tx_data = tx_helper.get_tx_data(search_id, actually_txt_all_successful_txs)
+        print(f"tx_data = {tx_data}")
+        tx_helper.clean_tx_data(tx_data)
         test_data = test_datas[i]
         if tx_data == test_data:
             print(f"Test {i} PASSED")
         else:
             print(f"Test {i} FAILED")
+
     print("ALL TESTS PASSED")
 
 
@@ -253,6 +284,7 @@ def main(args: argparse.Namespace):
 
     # Initialize the Web3 Manager
     w3_manager = Web3Manager(args.rpc_url)
+    strategy_mgr = StrategyManager
     w3 = w3_manager.w3
 
     multichain_addresses_path = os.path.normpath(
@@ -272,16 +304,16 @@ def main(args: argparse.Namespace):
     if args.task == "set_test_state":
         set_test_state_task(w3)
     elif args.task == "get_carbon_strategies_and_delete":
-        get_carbon_strategies_and_delete_task(w3, carbon_controller)
+        get_carbon_strategies_and_delete_task(w3, carbon_controller, args.from_block)
     elif args.task == "run_tests_on_mode":
-        test_strategies = get_test_strategies()
+        test_strategies = strategy_mgr.get_test_strategies()
         run_tests_on_mode_task(w3, carbon_controller, args.arb_mode, test_strategies)
     elif args.task == "run_results_crosscheck":
         run_results_crosscheck_task()
     elif args.task == "end_to_end":
         set_test_state_task(w3)
-        get_carbon_strategies_and_delete_task(w3, carbon_controller)
-        test_strategies = get_test_strategies()
+        get_carbon_strategies_and_delete_task(w3, carbon_controller, args.from_block)
+        test_strategies = strategy_mgr.get_test_strategies()
         run_tests_on_mode_task(args, w3, carbon_controller, test_strategies)
         run_results_crosscheck_task()
     else:
@@ -324,6 +356,19 @@ if __name__ == "__main__":
         choices=["single", "multi", "triangle", "multi_triangle"],
         help="Arbitrage mode to test",
     )
+    parser.add_argument(
+        "--timeout_minutes",
+        default=4,
+        type=int,
+        help="Timeout for the tests (in minutes)",
+    )
+    parser.add_argument(
+        "--from_block",
+        default=DEFAULT_FROM_BLOCK,
+        type=int,
+        help="Replay from block",
+    )
 
     args = parser.parse_args()
+    args.timeout = args.timeout_minutes * 60
     main(args)
