+++ strategy_created_df.csv
@@ -1,0 +1,5 @@
+,block_number,transaction_hash,owner,token0,token1,id,y0,z0,A0,B0,y1,z1,A1,B1
+0,19119009,0x15df2ddad07740d3370f981b2993cb67877ccfaf1889ba47b9f5ba2dc99df89c,0x28C6c06298d514Db089934071355E5743bf21d60,0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE,0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48,1701411834604692317316873037158841058240,1267460000029073000,1267460000029073000,0,4411844567835374,7634017977,7649999999,0,11433567603
+1,19119012,0x5c45e9657abd09fc93069ba112d6f7262111c7b3a47fbd4832af3a9c94e8abe0,0x28C6c06298d514Db089934071355E5743bf21d60,0x514910771AF9Ca656af840dff83E8264EcF986CA,0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE,20416942015256307807802476445906092688321,80000000000000000000,80000000000000000000,0,1399247190132272,0,0,0,17102934719008
+2,19119016,0x30e769285a60005a71b4ebdc1fe6c07eeedb3567b1d748e64a2cca42c4c3dd63,0x28C6c06298d514Db089934071355E5743bf21d60,0x6B175474E89094C44Da98b954EedeAC495271d0F,0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48,1361129467683753853853498429727072846786,3000000000000000000000,3000000000000000000000,0,6052452418541427,0,0,0,0
+3,19119019,0xa88f6a900d29ee8248f5eb1f6e2d56faf806e8d2a07d5c87494f913e3f0e8b76,0x28C6c06298d514Db089934071355E5743bf21d60,0x2260FAC5E5542a773Aa44fBCfeDf7C193bc2C599,0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE,9187623906865338513511114400657741710275,100000000,100000000,17694999,645747883,0,0,0,0
